ARM: i.MX: iim: Fix imx_iim_reg_write() return type
struct regmap_bus::reg_write should return int, not ssize_t.

+++ arch/arm/mach-imx/iim.c
@@ -291,7 +291,7 @@ err_out:
 	return ret;
 }
 
-static ssize_t imx_iim_reg_write(void *ctx, unsigned int reg, unsigned int val)
+static int imx_iim_reg_write(void *ctx, unsigned int reg, unsigned int val)
 {
 	struct iim_bank *bank = ctx;
